+++ pages/viewFormula.php
@@ -83,13 +83,20 @@ $(document).ready(function() {
 				$(row).find('td:eq(5)').addClass('alert-success');
 			}
 			
+			if(data.ingredient.classification == 4){
+				$(row).find('td:eq(0),td:eq(1),td:eq(5),td:eq(6)').addClass('alert-danger').append(' <i rel="tip" title="This material is prohibited" class="pv_point_gen fas fa-ban"></i></div>');
+            }
+			
 			if( data['usage_regulator'] == "IFRA" && parseFloat(data['usage_limit']) < parseFloat(data['final_concentration'])){
 				$(row).find('td:eq(6)').addClass('alert-danger').append(' <i rel="tip" title="Max usage: ' + data['usage_limit'] +'%" class="pv_point_gen fas fa-info-circle"></i></div>');
 			}else if( data['usage_regulator'] == "PV" && parseFloat(data['usage_limit']) < parseFloat(data['final_concentration'])){
 				$(row).find('td:eq(6)').addClass('alert-info').append(' <i rel="tip" title="Reccomended usage: ' + data['usage_limit'] +'%" class="pv_point_gen fas fa-info-circle"></i></div>');
+				
 			}else{
 				$(row).find('td:eq(6)').addClass('alert-success');
 			}
+			
+			
        },
 	   "drawCallback": function ( settings ) {
             var api = this.api();